Move notificaton functions back to ggrc module
This is the first step of moving all notification related function to
the notification package in the ggrc module.
