+++ src/ggrc/notifications/common.py
@@ -21,7 +21,7 @@ def getAppEngineEmail():
 
   Return the email of the user that will be set as sender. This email should be
   authorized to send emails from the server. For more details, see Application
-  Settins for email api authorized senders.
+  Settings for email api authorized senders.
 
   Returns:
     Valid email address if it's set in the app settings.
@@ -34,7 +34,7 @@ def send_email(user_email, subject, body):
   """Helper function for sending emails.
 
   Args:
-    user_email (string): Email of the recepient.
+    user_email (string): Email of the recipient.
     subject (string): Email subject.
     body (basestring): Html body of the email. it can contain unicode
       characters and will be sent as a html mime type.
@@ -54,3 +54,25 @@ def send_email(user_email, subject, body):
   message.html = body
 
   message.send()
+
+
+def modify_data(data):
+  """
+  for easier use in templates, it joins the due_in and due today fields
+  together
+  """
+
+  data["due_soon"] = {}
+  if "due_in" in data:
+    data["due_soon"].update(data["due_in"])
+  if "due_today" in data:
+    data["due_soon"].update(data["due_today"])
+
+  # combine "my_tasks" from multiple cycles
+  data["cycle_started_tasks"] = {}
+  if "cycle_started" in data:
+    for cycle in data["cycle_started"].values():
+      if "my_tasks" in cycle:
+        data["cycle_started_tasks"].update(cycle["my_tasks"])
+
+  return data
