+++ src/ggrc_workflows/views/__init__.py
@@ -24,28 +24,6 @@ env = Environment(loader=PackageLoader('ggrc', 'templates'))
 # module.
 
 
-def modify_data(data):
-  """
-  for easyer use in templates, it joins the due_in and due today fields
-  together
-  """
-
-  data["due_soon"] = {}
-  if "due_in" in data:
-    data["due_soon"].update(data["due_in"])
-  if "due_today" in data:
-    data["due_soon"].update(data["due_today"])
-
-  # combine "my_tasks" from multiple cycles
-  data["cycle_started_tasks"] = {}
-  if "cycle_started" in data:
-    for cycle in data["cycle_started"].values():
-      if "my_tasks" in cycle:
-        data["cycle_started_tasks"].update(cycle["my_tasks"])
-
-  return data
-
-
 def show_pending_notifications():
   if not permissions.is_admin():
     raise Forbidden()
@@ -53,7 +31,7 @@ def show_pending_notifications():
 
   for day, day_notif in notif_data.iteritems():
     for user_email, data in day_notif.iteritems():
-      data = modify_data(data)
+      data = common.modify_data(data)
   pending = env.get_template("notifications/view_pending_digest.html")
   return pending.render(data=sorted(notif_data.iteritems()))
 
@@ -63,29 +41,9 @@ def show_todays_digest_notifications():
     raise Forbidden()
   _, notif_data = notifications.get_todays_notifications()
   for user_email, data in notif_data.iteritems():
-    data = modify_data(data)
+    data = common.modify_data(data)
   todays = env.get_template("notifications/view_todays_digest.html")
   return todays.render(data=notif_data)
-
-
-def set_notification_sent_time(notif_list):
-  for notif in notif_list:
-    notif.sent_at = datetime.now()
-  db.session.commit()
-
-
-def send_todays_digest_notifications():
-  digest_template = env.get_template("notifications/email_digest.html")
-  notif_list, notif_data = notifications.get_todays_notifications()
-  sent_emails = []
-  subject = "gGRC daily digest for {}".format(date.today().strftime("%b %d"))
-  for user_email, data in notif_data.iteritems():
-    data = modify_data(data)
-    email_body = digest_template.render(digest=data)
-    common.send_email(user_email, subject, email_body)
-    sent_emails.append(user_email)
-  set_notification_sent_time(notif_list)
-  return "emails sent to: <br> {}".format("<br>".join(sent_emails))
 
 
 def _get_unstarted_workflows():
@@ -138,10 +96,6 @@ def init_extra_views(app):
   app.add_url_rule(
       "/_notifications/show_todays_digest", "show_todays_digest_notifications",
       view_func=login_required(show_todays_digest_notifications))
-
-  app.add_url_rule(
-      "/_notifications/send_todays_digest", "send_todays_digest_notifications",
-      view_func=send_todays_digest_notifications)
 
   app.add_url_rule(
       "/admin/unstarted_cycles",
