+++ test/integration/ggrc_workflows/notifications/test_one_time_wf_end_date_change.py
@@ -12,7 +12,6 @@ from mock import patch
 import os
 from ggrc import db, notifications
 from ggrc.models import Notification, Person
-from ggrc_workflows.views import send_todays_digest_notifications
 from ggrc_workflows.models import Cycle, CycleTaskGroupObjectTask
 from integration.ggrc_workflows.generator import WorkflowsGenerator
 from integration.ggrc.api_helper import Api
@@ -74,7 +73,7 @@ class TestOneTimeWfEndDateChange(TestCase):
       self.assertNotIn("due_today", notif_data[user.email])
 
     with freeze_time("2015-05-02 03:21:34"):
-      send_todays_digest_notifications()
+      notifications.send_todays_digest_notifications()
       _, notif_data = notifications.get_todays_notifications()
       self.assertEqual(notif_data, {})
 
@@ -89,7 +88,7 @@ class TestOneTimeWfEndDateChange(TestCase):
                        len(self.random_objects))
 
     with freeze_time("2015-05-04 03:21:34"):  # one day befor due date
-      send_todays_digest_notifications()
+      notifications.send_todays_digest_notifications()
       _, notif_data = notifications.get_todays_notifications()
       self.assertEqual(notif_data, {})
 
@@ -130,7 +129,7 @@ class TestOneTimeWfEndDateChange(TestCase):
       self.assertNotIn("due_today", notif_data[user.email])
 
     with freeze_time("2015-05-02 03:21:34"):
-      send_todays_digest_notifications()
+      notifications.send_todays_digest_notifications()
       _, notif_data = notifications.get_todays_notifications()
       self.assertEqual(notif_data, {})
 
@@ -188,7 +187,7 @@ class TestOneTimeWfEndDateChange(TestCase):
       self.wf_generator.activate_workflow(wf)
 
     with freeze_time("2015-05-02 03:21:34"):
-      send_todays_digest_notifications()
+      notifications.send_todays_digest_notifications()
       _, notif_data = notifications.get_todays_notifications()
       self.assertEqual(notif_data, {})
 
@@ -231,7 +230,7 @@ class TestOneTimeWfEndDateChange(TestCase):
       self.wf_generator.activate_workflow(wf)
 
     with freeze_time("2015-05-02 03:21:34"):
-      send_todays_digest_notifications()
+      notifications.send_todays_digest_notifications()
       _, notif_data = notifications.get_todays_notifications()
       self.assertEqual(notif_data, {})
 
@@ -260,7 +259,7 @@ class TestOneTimeWfEndDateChange(TestCase):
       self.assertIn("due_in", notif_data[user.email])
       self.assertEqual(len(notif_data[user.email]["due_today"]), 1)
 
-      send_todays_digest_notifications()
+      notifications.send_todays_digest_notifications()
 
     with freeze_time("2015-05-04 03:21:34"):  # due date
       user = get_person(self.user.id)
@@ -268,7 +267,7 @@ class TestOneTimeWfEndDateChange(TestCase):
       self.assertIn(user.email, notif_data)
       self.assertIn("due_today", notif_data[user.email])
       self.assertNotIn("due_in", notif_data[user.email])
-      send_todays_digest_notifications()
+      notifications.send_todays_digest_notifications()
 
     with freeze_time("2015-05-05 03:21:34"):  # due date
       _, notif_data = notifications.get_todays_notifications()
